IParser: Use long long int (#4046)
* Use long long int as the underlying integer data type in IParser.

* Extend number format to include `3e9`, `2.34e6`, `1'000'000`, etc.

+++ Src/Base/Parser/AMReX_IParser.cpp
@@ -54,7 +54,7 @@ IParser::operator bool () const
 }
 
 void
-IParser::setConstant (std::string const& name, int c)
+IParser::setConstant (std::string const& name, long long c)
 {
     if (m_data && m_data->m_iparser) {
         iparser_setconst(m_data->m_iparser, name.c_str(), c);
